Fix Spring bean configuration problem
The version of this project we forked is using a really old version
of Spring. The schema of the context file is not compatible with the
version that TeamCity 10.x provides at runtime (4.2.6). This change
updates the provided dependency to 4.2.6 and fixes the problem in the
context file (incorrectly specifying a singleton XML attribute).